feat(deps): angular version upgrade and converted into workspace (#65)
BREAKING CHANGE:
YES

## Description

This pull request introduces significant refactoring and modernization
of the `SearchComponent` in the `search-lib` package, focusing on
adopting Angular's new signal-based inputs/outputs, streamlining
configuration management, and improving testability. Additionally, it
includes workspace and dependency updates at the root level, and some
cleanup of outdated scripts and files.

Fixes #64 

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Intermediate change (work in progress)


## Checklist:

- [x] Performed a self-review of my own code
- [x] npm test passes on your machine
- [ ] New tests added or existing tests modified to cover all changes
- [ ] Code conforms with the style guide
- [ ] API Documentation in code was updated
- [ ] Any dependent changes have been merged and published in downstream
modules